+++ src/model.ts
@@ -58,11 +58,11 @@ ${setupMessage}`
     } else if (message.toUpperCase() === 'RESET') {
       return this.setupInfo(user)
     } else if (message.toUpperCase() === '?') {
-      return 'You can text me `CHECK` to see the delivery date, `RESET` to change your order number or email, and `STOP` to stop receiving texts from me.'
+      return 'You can text me CHECK to see the delivery date, RESET to change your order number or email, and STOP to stop receiving texts from me.'
     }
 
     // Otherwise, send a help message
-    return "Hello! I'm a Onewheel pinger bot to check on the delivery date so you don't have to. Type `?` to see a list of my commands."
+    return "Hello! I'm a Onewheel pinger bot to check on the delivery date so you don't have to. Type ? to see a list of my commands."
   }
 
   private async handleOrderNumber (user: UserObject, orderNumber: string) {
@@ -88,9 +88,9 @@ ${setupMessage}`
 
 I'll check daily, and keep you posted as to whether the date changes do that you don't have to.
 
-To see a list of commands which are available, text me \`?\`.
+To see a list of commands which are available, text me ?.
 
-If at any point you want to stop receiving messages, just text me \`STOP\` and I won't message you again.`
+If at any point you want to stop receiving messages, just text me STOP and I won't message you again.`
     } catch (err) {
       return err.message
     }